+++ src/domain/eventSearch/filterSummary.module.scss
@@ -1,0 +1,45 @@
+@import "colors";
+@import "fonts";
+@import "layout";
+@import "variables";
+
+.filterSummary {
+  .filtersWrapper {
+    padding: $basePadding;
+    background-color: $hel-white;
+    margin-bottom: 0.625rem;
+
+    @include respond_above(sm) {
+      margin-bottom: $largeMargin;
+    }
+
+    &:last-child {
+      margin-bottom: 0;
+    }
+
+    .titleFilterSummary {
+      font-size: $h4-font-size;
+      margin: 0 0 $baseMargin;
+    }
+
+    .clearButton {
+      display: block;
+      width: 100%;
+      margin-top: 1.5rem;
+      margin-bottom: -0.5rem;
+      padding: 0.5rem;
+      color: $hel-black-50;
+      border: 2px solid transparent;
+
+      &:hover {
+        border-color: $light-gray;
+        background-color: $lightest-gray;
+      }
+
+      &:focus {
+        outline: 2px solid $hel-black;
+        outline-offset: 0 0 1rem 0;
+      }
+    }
+  }
+}
