+++ src/domain/eventSearch/searchResultList.module.scss
@@ -5,8 +5,7 @@
 
 .searchResultListContainer {
   background-color: $hel-black-5;
-  padding-bottom: 10rem;
-  padding-top: 4.5rem;
+  padding: $largePadding 0;
 }
 
 .count {
@@ -27,51 +26,17 @@
   @include respond_above(sm) {
     grid-row-gap: unset;
     grid-column-gap: $largeMargin;
-    grid-template-columns: 2fr 5fr;
-  }
+    margin-left: auto;
+    margin-right: auto;
+    max-width: 68.75rem;
 
-  .filtersContainer {
-    .filtersWrapper {
-      padding: $basePadding;
-      background-color: $hel-white;
-      margin-bottom: 0.625rem;
-
-      @include respond_above(sm) {
-        margin-bottom: $largeMargin;
-      }
-
-      &:last-child {
-        margin-bottom: 0;
-      }
-  
-      .titleFilterSummary {
-        font-size: $h4-font-size;
-        margin: 0 0 $baseMargin;
-      }
-
-      .clearButton {
-        display: block;
-        width: 100%;
-        margin-top: 1.5rem;
-        margin-bottom: -0.5rem;
-        padding: 0.5rem;
-        color: $hel-black-50;
-        border: 2px solid transparent;
-
-        &:hover {
-          border-color: $light-gray;
-          background-color: $lightest-gray;
-        }
-
-        &:focus {
-          outline: 2px solid $hel-black;
-          outline-offset: 0 0 1rem 0;
-        }
-      }
+    &.hasFilterSummary {
+      grid-template-columns: 2fr 5fr;
+      margin-left: unset;
+      margin-right: unset;
+      max-width: unset;
     }
   }
-
-  
 
   .eventListWrapper {
     .loadMoreWrapper {
